requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.0
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.3
distro==1.9.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
durationpy==0.10
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==1.0.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==1.0.3
langchain-core==1.0.2
langchain-google-genai==3.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.39
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
numpy==2.3.4
oauthlib==3.3.1
onnxruntime==1.23.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.11.0
overrides==7.7.0
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
posthog==5.4.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyMuPDF==1.26.5
pyparsing==3.2.5
pypdfium2==5.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.38.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
